Added support for ARC.
Weak references are still not supported, but code that doesn't use them (i.e. any code that wants to be compatible with OS X 10.6) will work fine.

The current implementation is VERY inefficient and has a large number of missed optimisation opportunities: this is the 'make it right' phase, and should be almost equivalent to explicit retain / release code.



git-svn-id: http://svn.gna.org/svn/gnustep/libs/libobjc2/trunk@33397 72102866-910b-0410-8b05-ffd578937521